#9874: Merge log_sigmoid_bw to TTNN
