+++ docs/source/ttnn/ttnn/api.rst
@@ -177,6 +177,7 @@ Pointwise Unary
    ttnn/sub_bw
    ttnn/frac_bw
    ttnn/trunc_bw
+   ttnn/log_sigmoid_bw
 
 Pointwise Binary
 ================
