+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -914,8 +914,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.square_bw
 
-.. autofunction:: tt_lib.tensor.log_sigmoid_bw
-
 .. autofunction:: tt_lib.tensor.tanhshrink_bw
 
 .. autofunction:: tt_lib.tensor.threshold_bw
