+++ docs/source/ttnn/ttnn/ttnn/log_sigmoid_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.log_sigmoid_bw:
+
+ttnn.log_sigmoid_bw
+###################
+
+.. autofunction:: ttnn.log_sigmoid_bw
