+++ tt_eager/tt_dnn/op_library/backward/backward_ops.cpp
@@ -1458,28 +1458,6 @@ std::vector<Tensor> square_bw(const Tensor& grad, const Tensor& input, const Mem
     return operation::decorate_as_composite(__func__, _square_bw)(grad, input, output_mem_config);
 }
 
-// return: grad_output * (max_deriv - sign * (z / (1 + z)))
-// z = exp(-abs(input))
-std::vector<Tensor> _log_sigmoid_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    Tensor max_deriv = where(ttnn::ltz(input, output_mem_config), 1, 0, output_mem_config);
-    Tensor in_sign = where(ttnn::ltz(input, output_mem_config), 1, -1, output_mem_config);
-    Tensor in_abs = ttnn::abs(input, output_mem_config);
-    Tensor z = ttnn::exp(ttnn::neg(in_abs, output_mem_config), false, output_mem_config);
-
-    Tensor mul_z = ttnn::multiply(z, ttnn::reciprocal((ttnn::add(z, 1.0f, std::nullopt, output_mem_config)), output_mem_config), std::nullopt, output_mem_config);
-
-    Tensor mul_sign = ttnn::multiply(in_sign, mul_z, std::nullopt, output_mem_config);
-    Tensor sub_max = ttnn::subtract(max_deriv, mul_sign, std::nullopt, output_mem_config);
-
-    Tensor grad_result = ttnn::multiply(grad, sub_max, std::nullopt, output_mem_config);
-    grad_tensor.emplace_back(grad_result);
-    return grad_tensor;
-}
-std::vector<Tensor> log_sigmoid_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _log_sigmoid_bw)(grad, input, output_mem_config);
-}
-
 // tanhshrink
 // result:  torch.square(torch.tanh(input)) * grad_data
 std::vector<Tensor> _tanhshrink_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
