+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -24,6 +24,7 @@ enum class UnaryBackwardOpType {
     SUB_BW,
     FRAC_BW,
     TRUNC_BW,
+    LOG_SIGMOID_BW,
 };
 
 struct UnaryBackwardFunction{
